+++ frontend/package.json
@@ -55,7 +55,7 @@
     "geist": "latest",
     "input-otp": "1.4.1",
     "lucide-react": "^0.454.0",
-    "next": "^15.5.4",
+    "next": "16.0.1",
     "next-intl": "^4.3.9",
     "next-themes": "^0.4.6",
     "react": "^19.1.1",
