+++ frontend/src/core/lib/firebase-client.ts
