--- frontend/src/core/services/api-client.ts
@@ -1,32 +1,0 @@
-﻿import axios from "axios";
-import { API_BASE_URL, API_TIMEOUT } from "@/core/config/api";
-
-export const apiClient = axios.create({
-	baseURL: API_BASE_URL,
-	timeout: API_TIMEOUT,
-	withCredentials: true,
-	headers: {
-		"Content-Type": "application/json",
-	},
-});
-
-export const setAuthToken = (token: string | null) => {
-	if (!token) {
-		delete apiClient.defaults.headers.common.Authorization;
-		return;
-	}
-
-	apiClient.defaults.headers.common.Authorization = `Bearer ${token}`;
-};
-
-apiClient.interceptors.response.use(
-	(response) => response,
-	(error) => {
-		if (axios.isAxiosError(error) && error.response) {
-			const message = error.response.data?.message ?? error.message;
-			return Promise.reject(new Error(message));
-		}
-
-		return Promise.reject(error);
-	},
-);
