+++ frontend/src/modules/scholarships/services/scholarship.interface.ts
@@ -1,0 +1,10 @@
+import type { ScholarshipSearchParams, ScholarshipSearchResponse } from "../types";
+
+export interface IScholarshipSearchService {
+  /**
+   * Search scholarships using full-text search
+   */
+  getScholarships(
+    params: ScholarshipSearchParams,
+  ): Promise<ScholarshipSearchResponse>;
+}
